Some fixes for data types

+++ pom-root/interfaces/src/main/java/ar/edu/itba/paw/interfaces/PropertyDao.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface PropertyDao {
     public Property findById(final long id);
     public List<Property> getAll();
-    public Long createProperty(String street, Integer number, Integer floor, String apartment, Property.Type type, Integer userId, Long price, Integer coveredArea, Integer totalArea, Integer rooms, Integer baths, Boolean garage, Integer taxPrice);
-    public List<Property> getFavorites(Integer userId);
+    public Long createProperty(String street, Integer number, Integer floor, String apartment, Property.Type type, Long userId, Long price, Integer coveredArea, Integer totalArea, Integer rooms, Integer baths, Boolean garage, Integer taxPrice);
+    public List<Property> getFavorites(Long userId);
 }
